Namespace some code
